Add GlobalTask::wakeUp, remove unnecessary snooze() calls
Add a method to wake up a task immediatatly -
GlobalTask::wakeUp().

This is logically identical to calling GlobalTask::snooze() with the
argument 0, but (IMHO) is more self-documenting.

Additionally, remove unnecessary calls to snooze(0) - a task is
rescheduled to run immediately by default if it returns true from
run(), it is not necessary to call snooze(0) explicitly.

Change-Id: If70f77cecfe7c085b901c871fa9eb37b5751fa52