MB-24765: Split ephemeral tombstone purger into two tasks
Split the existing EphTombstonePurgerTask in two:

1. EphTombstoneHTCleaner task which is responsible for marking old
   tombstones in the HashTable as stale (and removing from HashTable).

2. EphTombstoneStaleItemDeleter task which is responsible for walking
   the SequenceLists and removing stale items.

The reason for this change is that currently the combined
EphTombstonePurgerTask can run for a long time, consuming a NonIO
thread and blocking other tasks from running. To solve this we want to
make purging incremental; so only a small amount of work is done at a
time - similar to the Defragmenter task. To achieve this however the
purging activities need to be migrated to a pause/resume visitor
model, and each iteration (HashTable and SequenceList) will need a
seperate visitor. As such it will be simpler (and easier to measure
their respective runtimes) if the purging work is split into two
tasks.

Change-Id: Ifa422ad809ea8e8282fbba240d1b04140f2c78d6