MB-24765: Change PauseResumeVBVisitor to take VBucket
Modify PauseResumeVBVisitor::visit() to take a VBucket reference
instead of vbid + HashTable reference. This allows instances of
PauseResumeVBVisitor to know the actual VBucket which is being
visited, not just the underlying HashTable.

This is necessary for subsequent changes to convert Ephemeral
tombstone purger to pause/resume, as it needs to have access to the
VBucket to move items.

Change-Id: I2086c66d716beee2c7ac265c0f09cf714c664cc7