Fix incorrect hash bucket in pauseResumeVisit
In HashTable::pauseResumeVisit() we are incorrectly using the lock
number (e.g. 0..47) instead of the HashTable bucket number. The effect
of this is that the HashBucketLock we pass into the visit() method has
an incorrect hash bucket number.

Fortunately there is only one existing user of pauseResumeVisit
(DefragmenterTask), and this doesn't need to mutate the StoredValue so
this doesn't have any effect /currently/. However it does cause
problems if one attempts to modify the StoredValue - for example when
the Ephemeral HTTombstonePurger is converted to use pause/resume.

Change-Id: I4166de8a0e28b92fd27962065f7ef0e23bf4432e