MB-24861: Remove UUID:0 from vbucket_state::reset()
When resetting a vbucket_state, don't add a failover table entry with
a UUID of zero.

While it's been hard to reproduce, this could in theory cause another
instance of UUID becoming zero on disk - vbucket_state::reset() is
sometimes called during a Flush operation, and in which case the UUID
can end up as zero.

Note that due to the fix for MB-20822, we sanitize UUIDs when a
FailoverTable is created, and hence while we _could_ have a UUID of
zero on disk it will be corrected when that is loaded into memory.

This is a speculative commit, addressing an issue seen during code
review - we haven't actually seen this hit "in the wild", but it can
be forced to show up in unit tests.

Change-Id: I9ab21d6fcc1594e60ce465d650b7d685743f94a4