MB-25065: Estimate mem req for new item correctly in Ephemeral
We estimate the memory usage increase from the addition of a new
item to the in-memory data structures, to ensure that the 'mem_used'
is less than the 'max_size'. For that, in Ephemeral buckets we must
use the size of 'OrderedStoredValue' than the size of 'StoredValue'.

This commit (i) Uses OrderedStoredValue in Ephemeral buckets
(ii) Refactors the code related to the memory usage estimation.

Change-Id: I789060f83f1cd07892ecca4dc0de7087a66c322b