MB-25105: Fix bug in ExpiredItemsDeletedFirst test
There is a bug in the validation of the postconditions of this test -
when checking to see if various items exist or not in the vBucket, the
test uses the public get() method - which will delete any items whose
TTL has been exceeded on access.

This is problematic as it defeats the point of the test - to verify that
items with a TTL which has expired are removed by the Item pager.

Fix by just checking the item count instead of the specific items.

Change-Id: I2703025e325fd0f243d99895ccab19337565954b