MB-24860: Destruct ThreadLocals on object destruction or thread exit
We need to ensure that thread-local variables are destructed correctly
to prevent memory leaks. Unlike other variable types, thread-locals
can be considered to have two points when they should be deleted - (1)
when the associated thread exits, and (2) when the owning object is
detroyed.

A previous patch (8bccd7189a) has adddressed (1), however we still
have the scenario where the owning object is destroyed but not the
thread - from the MB:

    There's a secondary leak that was introduced at the same time (but
    wasn't covered by our tests at the time).

    It's similar to this leak as fully described in
    http://review.couchbase.org/#/c/82535/ , in effect the memcached
    worker threads get shutdown after the bucket has been destructed
    (and therefore after the ThreadLocal object has been destructed
    and the pthread specific data has been deleted) which means the
    per-thread destructor for the memory counters doesn't get run on
    those threads.

    This affects memcached in normal use (not just tests) and results
    in 16 bytes per front-end thread (which for some customers is a
    lot of threads) leaked every time a bucket is shutdown.

To solve this we need to ensure that /either/ of the above events -
containing object destruction or thread exit - cause the thread-local
to be deleted.

Change-Id: If500aabd9ab082f2beb9c6d0483596b7a42ac7ae