MB-28941: Tracing: Allow specifying begin/end time
The Tracer begin() and end() methods implicitly use the current time
(ProcessClock::now()) as the stat/end time. However we may already
have the start and/or end time to use. Given ProcessClock::now() is
generally not zero cost; add a start/endTime parameter (defaulted to
ProcessClock::now) to begin() and end() which allow the caller to
specify the time_point to use.

Make use of this for the overall Request duration, as we /do/ have
both the start & end time for them.

Doesn't have any direct effect on the microbenchmarks (as they don't
otherwise have a timestamp to use).

Change-Id: I10fc84b65cf924383a785c596e001d60944e589d
Reviewed-on: http://review.couchbase.org/92302
Reviewed-by: Trond Norbye <trond.norbye@gmail.com>
Tested-by: Build Bot <build@couchbase.com>