MB-28941: Remove duplicate clock reads when tracing
Currently a number of the Session Tracing events essentially duplicate
existing timed durations; however they perform their own clock
reads. This is both inefficient (reading the clock costs cycles;
particulary on virtualized systems), and inprecise - we cant' easily
compare the timestamps in the histograms with the cookie's timings as
they are timing slightly different durations.

To address this, make use of the new ScopeTimer RAII classes to time
the start & end just once, but broadcast those time values to both the
per-bucket historgrams and per-Cookie Tracing.

Note that as part of this we no longer distinguish between EvpStore
and EvpStoreIf in the Session Tracing, as one is implemented in terms
of the other (we didn't have seperate histograms for them). Splitting
them apart was probably more detail than users care about anyway, so I
don't think this is a problem.

Change-Id: I50f87a99ce74fadb61931f1b44011e42207c5020
Reviewed-on: http://review.couchbase.org/92324
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@gmail.com>