MB-28941: Remove TraceCodes which we didn't previously track in histograms
Session Tracing is currently exceeding its performance budget of 1%
overhead. As such, remove all the TraceCodes which were not previously
tracked in histograms prior to Session Tracing - remove the calls to
TRACE_SCOPE(code), along with the enum definitions.

Also remove codes which were defined but never used
(e.g. SETITEMINFO).

In micro-benchmarks this reduces the cost of our simulated SET
operation by 1/3 (as we've reduced the number of trace spans from 3 to
2).

Change-Id: I530eb247b9400ea0783adc0a848d071bb5a7aa6b
Reviewed-on: http://review.couchbase.org/92326
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@gmail.com>