MB-28941: Make SET_WITH_META tracing cheaper by reusing histogram times
For SET_WITH_META operations, instead of using the TRACE_SCOPE macro
(and generating it's own begin/end time), make use of the existing
timestamps which are used for setWithMetaHisto. This has two benefits:

1. The timestamps in session tracing and the histograms are
   consistent.

2. We save 2x calls to ProcessClock::now for SET_WITH_META ops.

To achieve this, we need the ability to record session tracing begin
and end events independently, as the timing for SET_WITH_META isn't
limited to a single block (due to EWOULDBLOCK handling). As such,
introduce two new macros for this purpose:

* TRACE_BEGIN(cookie, code [,timestamp]) - Begins a trace span for the
  given cookie and tracecode.

* TRACE_END(cookie, code [,timestamp]) - End a trace span for the
  given cookie and tracecode.

In both instances, the caller can optionally specify the timestamp to
record; if not specified the current time is used.

Change-Id: Iccc3ae3e36b9b9b51e161b18b0288a6f9e05d2df
Reviewed-on: http://review.couchbase.org/92489
Reviewed-by: Paolo Cocchi <paolo.cocchi@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Jim Walker <jim@couchbase.com>