MB-33186: Add KVStore::getRange()
Add a new method to KVStore - getRange():

    Searches the given vBucket for all items with keys in the half-open
    range [startKey,endKey). For each item found invokes the given callback.

This is to be used by warmup, to locate all prepared SyncWrites which
have been persisted to disk but not yet committed to disk. To do this,
Warmup needs to perform a range scan on the byKey index, fetching any
documents which start with the DurabilityPrepare namespace (0x2).

Implemented for CouchKVStore and RocksDBKVStore as they both already
support range scans on the key index.

Change-Id: I839e4f4699f6c9e65b6f5a93a9094558509b9244