HashTable::unlocked_find: Don't assume ordering of prepared and committed
HashTable currently assumes that prepared SyncWrites will always be in
the hash bucket chain ahead of any previously committed StoredValue
(for the same key).

While this is the case in normal operation (as a prepare happens
temporally after any previous commit, and hence will be inserted into
the head of the hash chain), it is not the case during warmup, where
we will load prepared SyncWrites first.

As such, make HashTable::unlocked_find() robust to the exact hash
chain order, adding unit tests to check this.

(The alternative would be to add an invariant that matching keys are
always stored in descending seqno order in the HashTable).

Change-Id: I0d043236fd22edd74764e468f00e0bb710eb8776