MB-36723 [1/2] Change default ExecutorPool #reader / #writer to #cores
Summary
-------

Change the calculation of how many Writer threads are created to be
equal to the number of CPU cores.  The calculation of how many Reader
threads to create is currently dependent on how many Writers are
created, so also change them to default to the number of CPU cores.

This shouldn't affect single-bucket tests yet; until the number of
shards is changed these extra threads will just be idle. See the next
patch in this series for the num_shards increase.

Background
----------

The rationale for this is that the current throughput of
persistence-dependent ops (e.g. SyncWrites) with the current defaults
(4 shards / 4 writer threads) is generally much lower than modern
SSD-based systems are capable of.

Given couchstore's disk IO is all synchronous (blocking), the number
of concurrent IO operations issued to the disk subsystem is limited to
the number of Flusher tasks (i.e. number of shards) - default 4.

Commodity SSDs are typically capable of 32 concurrent IOPs (queue
depth), enterprise-grade SSDS, mutli-disk RAID and NVMe disks are
capable of _significantly_ more; so only issuing up to 4 writes ops
concurrently is insufficient to fully utilise disks.

Note: Reader / Writer threads are typically IO-bound, so a Writer
thread should only consume a fraction of a CPU core. As such, creating
one per CPU core shouldn't cause too much contention with other,
CPU-bound threads in the system.

Note (2): For maximum IO throughput we should ideally create as many
Reader / Writer threads (and shards) as concurrent IOPs the system can
support (given we use synchronous IO).  However, knowing the number of
concurrent IOPs the system can support is hard, so we use #CPUs as a
sensible default proxy for it - machines with lots of CPU cores are
more likely to have more IO than little machines.

Change-Id: I41ef97154b05c33f88f00467556b1fa5937e5f9b