MB-36723 [2/2] Auto-set num_shards based on #cores
This is the second patch in the series to improve disk write
throughput, by increasing the number of Flusher tasks which can run
concurrently.

It changes the default number of shards - from a fixed value (4) to
instead auto-configure based on the number of CPU cores in the
system. Specifically, max_num_shards is set to the highest power-of-2
not greater than the CPU count. For example:

    CPUs  max_num_shards
    1     1
    2     2
    3     2
    4     4
    ...
    7     4
    8     8

Recall that one Flusher and BGFetch task is created per shard, so
increasing the number of shards increases the number of tasks which
can _potentially_ be executed parallel. By also increasing the default
number of Reader / Writer threads based on the CPU count (see patch 1)
we can actually run more flusher work in parallel.

Change-Id: I20a5c74349c87a6abd805d4632a32cfcddaa1bf0