MB-42029: FollyExecPool: Simplify TaskProxy ownership
During investigation of this MB, it was noted that the TaskProxy no
longer needs to use shared ownership (shared between IO thread pool
and CPU thread pool). We must only destroy a TaskProxy from an IO
thread (to avoid any data races given the IO thread created the object
initially), and indeed that is already handled by
removeTaskAfterRun().

As such, we can simplify the owenership model and just use a
unique_ptr for TaskProxy instead of shared_ptr - passing a plain
reference to TaskProxy on CPU thread pool.

Change-Id: I2d09f8febf63802bcaadd6e9d02081736b11d563