MB-52311: [7/n] Pause / Resume Bucket: Add cancellation support
If a Resume() request is made while a previous Pause() request is
still executing, we don't want to have to wait until the Pause()
completes before resuming. This is particulary important when
considering failover - if ns_server decides it needs to failover a
node it needs to being the Bucket back to the Ready state asap so it
can examine and/or set vBucket state to perform the failover.

As such, we want to be able to cancel a running Pause() operation if a
Resume() request arrives.

This patch adds cancellation support in the daemon-level handling of
pause() - e.g. check for cancellation while waiting for clients to
disconnect. A follow-up patch will expand checking for cancellation
inside ep-engine (which also has some potentially slow-running tasks
as part of pause).

=== Implementation Details ===

We do this using Cancellation Tokens - which Folly provides an
implementation. Quoting from folly/CancellationToken.h:

    A CancellationToken is an object that can be passed into an
    function or operation that allows the caller to later request that
    the operation be cancelled.

    A CancellationToken object can be obtained by calling the
    .getToken() method on a CancellationSource or by copying another
    CancellationToken object. All CancellationToken objects obtained
    from the same original CancellationSource object all reference the
    same underlying cancellation state and will all be cancelled
    together.

    If your function needs to be cancellable but does not need to
    request cancellation then you should take a CancellationToken as a
    parameter.  If your function needs to be able to request
    cancellation then you should instead take a CancellationSource as
    a parameter.

In terms of how this maps to kv_engine, we add a CancellationSource to
each Bucket object, and whenever a Pause() operation occurs we:

1. Construct a CancellationSource for this pause request.

2. Constuct a CancellationToken from the CancellationSource and pass
   the token into the BucketManager (and later further down into the
   engine).

At suitable points in the pause process (typically before some
long-lived / blocking operation) the CancellationToken can be checked
to see if cancellation has been requested, if so then we cancel the
pause, undoing any steps in the Pause process which have already been
performed.

Note that one can also register CancellationCallbacks with the
CancellationToken (multiple callbacks can be registered). Each
callback can perform whatever necessary "undo" logic to revert the
back to the unpaused state.

To trigger cancellation, the Resume() method checks to see if
Cancellation is possible (i.e. a Pause request is in progress), and if
so requests cancellation and returns early - the cancelled Pause()
request is responsible for undoing any pause steps already executed.

=== Locking ===

The CancellationSource object guarded with the Bucket::mutex, for two
reasons:

1. We need to serialise the creation and reset of CancellationSource
   when a Pause request starts and finishes (i.e. resume needs to
   atomically check if a pause is currently in-flight and cancel that
   instead of performing the normal resume() path).

2. When cancellation is triggered we need to run at least one callback
   under the Bucket::mutex as we must change the Buckets' state back
   to Ready. Given this callback can be triggered on the cancelling
   thread (more commonly) *or* inline on the pausing thread when
   setting up the cancellation callback, we need to hold mutexes in a
   consistent way.

Change-Id: I1125b6d96d215989b975eda3a4fa81837cd9ea5f